// src/modules/formValidation/enum/wordpressFormValidation.enum.ts
export enum WordpressFormValidationEnum {
    VALIDATION_PASSED = "",
    FIELD_REQUIRED = "100000",
    DATA_LENGTH_INVALID = "100001",
    DATA_NOT_MATCH_REGEXP = "100002",
    REGEXP_INVALID = "100003",
    INVALID_SELECT_OPTION = "100004",
    NUMBER_NOT_IN_VALID_RANGE = "100005",
    INVALID_CONFIRM_PASSWORD = "100006",
    FIELD_COMPLEX_REQUIRED = "100007",
    FIELD_COMPLEX_REQUIRED_ALL = "100008",
    DATE_INVALID_FORMAT = "100009",
    DATE_NOT_IN_VALID_RANGE = "100010",
    TIME_INVALID_FORMAT = "100011",
    FAIR_INVALID = "100012",
    INVALID_ACCEPTANCE_OPTION = "600001",
    FILE_PATH_EXCEED_LIMIT = "700001",
    FILE_MISSING = "700002",
    FILE_INVALID_FILE_TYPE = "700003",
    FILE_INVALID_FILE_SIZE = "700004",
    FILE_FAIL_TO_VALIDATE = "790000",
    CAPTCHA_MISSING = "800001",
    CAPTCHA_INVALID = "800002",
    X_FORWARDED_FOR_MISSING = "800003",
    FIELD_TYPE_NOT_FOUND = "900000",
    GENERIC_FIELD_TYPE_DEFINITION_NOT_IMPLEMENTED = "900001",
    DEDICATE_FIELD_TYPE_DEFINITION_NOT_IMPLEMENTED = "900002"
}